pyproject.toml
[tool.poetry]
name = "providers"
version = "0.1.0"
description = ""
authors = ["Sivan Becker <sivanbecker@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
flask-marshmallow = "^0.14.0"
Flask-RESTful = "^0.3.8"
flask_sqlalchemy = "^2.4.4"
marshmallow-sqlalchemy = "^0.23.1"
Flask-SQLAlchemy = "^2.4.4"
psycopg2 = "^2.8.6"
logbook = "^1.5.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pudb = "^2019.2"
mypy = "^0.790"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
